Unskip some tests
